src/requirements.txt
duckdb>=0.10.0
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=12.0.0
statsmodels>=0.14.0
matplotlib>=3.7.0
pyyaml>=6.0
